git-svn, perl/Git.pm: extend Git::prompt helper for querying users
git-svn reads usernames and other user queries from an interactive
terminal. This cause GUIs to hang waiting forever for git-svn to complete
(http://code.google.com/p/tortoisegit/issues/detail?id=967) if they are
not connected to the terminal. Give *_ASKPASS helpers preference over
terminal when asking for usernames in git-svn to work around this issue.

This change is technically a small regression for users who have *_ASKPASS
environment defined and expect the username to be read from the terminal;
they now have to type their username to GUI helpers that are likely to be
designed to read passwords without echoing. We may want to swap the
precedence so that we try terminal first when we can to fix this, but that
is outside the scope of this patch.

Signed-off-by: Sven Strickroth <email@cs-ware.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>