rev-list test: show that --not does not handle time skew
The internal propagation of UNINTERESTING marks during the revision
walk processes commits in date-guided search order[*], just like the
normal commit walk.  This can occasionally break the commit filtering
quite badly, as shown by this test.

We confuse the walker by having a short chain of commits that are in
the past, relative to the base they are built on.  If we then ask for
the past commits to be excluded from listing, the UNINTERESTING marks
on them will not be propagated until it is too late.

A short chain is needed because the walker scans a few more commits in
an attempt to cope with such broken timestamps.  The minimum chain
length for which this test works is -- empirically -- 7.

[*] It keeps a list of "current" commits, sorted by date.  At every
step it processes the newest commit from the list, and adds all of its
parents to the list (if not seen so far).  Note that this order does
not have any "children before parents" guarantees.

Signed-off-by: Thomas Rast <trast@student.ethz.ch>