gitk: split out diff part in $commitinfo
So far we just parsed everything after the headers into the "comment"
bit of $commitinfo, including notes and -- if you gave weird options
-- the diff.

Split out the diff, if any, into a separate field.  It's easy to
recognize, since the log message is indented but the /^diff / that
starts a diff is not.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>