gitk: support unstuck arguments for -S and -G
gitk has traditionally only supported the stuck forms of options,
except for -n.  However, with improved parsers in core git, most
options can be given with either stuck or unstuck arguments.

This change only touches the low-hanging fruit of -S and -G, which can
be handled with minimal fuss using the existing logic.  For some
strange reason they are also the only gitk-supported options that my
fingers spell in the unstuck form.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>